refactor: migration に型定義を合わせる

+++ src/main/java/com/example/online_workspace/repositories/RoomRepository.java
@@ -72,7 +72,7 @@ public interface RoomRepository {
 		String categoryDescription,
 		int categorySortOrder,
 		String workStyle,
-		int maxMembers,
+		short maxMembers,
 		String visibility,
 		String status,
 		long creatorId,
